+++ tests/testthat/test-scale-roi.R
@@ -1,20 +1,20 @@
 context("test-scale-roi")
 
-test_that("scale_roi works", {
+test_that("scale_layer works", {
 	expect_error(
-		scale_roi(layer = NULL, roi = densitygrid),
+		scale_layer(layer = NULL, roi = densitygrid),
 		'layer must be provided and expected type is raster.'
 	)
 
 
 	expect_error(
-		scale_roi(layer = dem, roi = NULL),
+		scale_layer(layer = dem, roi = NULL),
 		'roi must be provided.'
 	)
 
 
 	expect_error(
-		scale_roi(layer = 'a', roi = densitygrid),
+		scale_layer(layer = 'a', roi = densitygrid),
 		'layer must be provided and expected type is raster.'
 	)